Update missed tsuite model to Fe+ of Smyth+19
slow/feii_blr_n11_p20.in was not properly updated on the FeII branch.

The file contained a blank line before the monitored values updated here
which prevented them from being exercised.  The blank line was removed
on master, so when that was merged on FeII, the monitors became active
with their outdated values.