Update tsuite to temper FLT_IS_DBL botches
Several sims in slow/ botch in all-double runs with the new default for
the Fe+ model (Smyth+19).  A botch in grid_extreme.in has migrated after
Gargi Shaw's update of UMIST rates to 2012.

Reset the monitored values associated with these botches to the mean of
the min/max values.  Also, slightly widen error margins in a couple of
occasions to allow for successful completion of sims across the entire
test suite.