Fix how ExtraLymanLinesJxx are indexed
The total number of levels included in the stack of states stJxx is the
sum of nLyman_alloc and numlevels_max.
All indices that loop over a principle quantum number is named nHi,
whereas ipHi is the label for the index of a level, to reduce ambiguity
in indexing variables.
The first two elements in ExtraLymanLinesJxx are junked, so that the
elements of ExtraLymanLinesJxx begin with nHi=2.
A new variable was created nLymanHLike to carry the number of extra
lyman lines for H-like ions, so that it needs be set only once in
iso_create.cpp.
Setting the states, junking states, setting the Hi and Lo states and
AddLine2Stack for the ExtraLymanLinesJxx are now all being done in a
new procedure called iso_assign_extralyman_levels. Because QN2Index
required to find ipHi, so needs to be done after iso_assign_quantum_numbers.