Add URL to Cloudy paper in README.md
Previously, the Cloudy reference was contained only in the file
data/citation_cloudy.txt.  Gary suggested we add a URL somewhere to ease
access to the most recent review/release paper.  The top README.md file
is the most obvious spot to do so.

Note that currently both the URL and the citation in the file refer to
version C17.  (An update to the C23 reference is imminent.)