Ignore trailing 'abstract' from ADS refs in Stout
With the new UI on ADS, a copy-paste of the URL into Stout introduces
a trailing 'abstract'.  This confuses the script that extracts Stout
references, which then queries ADS with 'abstract' as the bibcode.

Fix script to ignore that word, so as to obtain the actual bibcode.

Update the JSON file of Stout references, to correctly process Fe+.
Fetch the Smyth+2019 reference in the process.