Introduce datasets JSON stout refs file
The scripts and JSON data file that handle Stout references are not
designed to handle multiple datasets per atomic model, as is the case
for Fe+.

This is the first step toward accommodating present data needs in Stout.

Convert the current contents of the 'ref' entry in JSON and Perl data
structures into the contents of a 'default' dataset object, which is
what 'ref' now holds.