Fix bug with number of lines per Stout refs table
Previously, a bug existed that caused some (continued) tables to write
in the page footer:

The variable that keeps track of the total number of lines per table was
reset to 0, instead of to the number of lines of the species written to
the table right after the split.

This also permits longer tables with prep-bib.sh.