Exit gracefully sims of illuminated cooling gas
In calculations of a gas parcel cooling in the presence of intense
ionizing radiation, the gas may reach a temperature floor higher than
the target temperature (uses as a stopping criterion).

For instance, the sim may be set up to let the gas cool down to, say,
10,000 K, but due heating by ionizing radiation, it may not be able to
cool below 100,000 K.

Previously, the code would keep iterating until a stopping criterion on
the number of iterations was met, or, worse, until the code would crash.

Modify this behavior: Gracefully exit the sim when a temperature floor
is reached.

The floor is identified when the temperature stays constant to within
some tolerance for the last N timesteps.

Cloudy then declares it has converged, even though the requested number
of iterations may not have been reached.