Add 'stop time' command
In cooling calculations of an illuminated gas, it may be desired to
integrate to a given total time, instead of to a temperature limit.

Add command 'stop time' to do so.

Model it after the homologous 'stop time when temperature' command.
This is preferred over modifying the 'time' command, which expects
the intended total elapsed time as its second argument, because that
command is intended for fluctuating continua.