Change release number policy
Christophe Morisset pointed out that the Cloudy version that is printed
on output differs depending on whether the source was obtained as a
tarball, or was git-cloned.

In the tarball version, the version is compiled from the major/minor
release numbers set in version.cpp, while in the cloned version, it is
obtained from gitversion.sh.

Modify the latter to produce the same string as the internal code (e.g.,
23.01), and drop the '-modified' string to not disrupt downstream code
like pyCloudy.

Modify version.cpp to anticipate the release number format, and get rid
of the release branch name, as it is not set on releases.

On development branches the version shown on output has not changed.