Fix bug with isotropic illumination in igm_z3.in
Previously, the use of the command 'illumination isotropic' with the
blackbody radiation field in the named sim led to a crash.  That was
because while for the total intensity the value corrected for isotropic
continua was used, for the inward component the uncorrected value was
used.  With the former nearly extincted, the inward was found to exceed
the total, hence the crash.

Use the corrected inward emissivity when adding the line to the stack.

Note that the sim is not modified.  These changes are in anticipation
of changes to the behavior of the 'illumination' command.