Add method for debugging blends on the line stack
Using the command 'illumination isotropic' with the sim
limit_lte_he1_nomole_ste_nocoll.in leads to a bug, because the blend
selected as normalization line has zero intensity.

While investigating the source of this behavior, it proved useful to be
able to inspect the individual blend components.  The new method enables
that.

Note that the sim is not modified.  A new normalization line should be
selected for isotropic illumination.