Fix the ELEMENT <NAME> TABLE command
The code allowed the ELEMENT <NAME> TABLE command to be used for multiple
elements, however due to an oversight the size of those tables needed to be
identical. This commit fixes that and additionally cleans up a lot of
duplicated code.

In commit fa14275f a modified debug statement in atmdat_chianti.cpp was
overlooked. That change is reverted here.