This commit is extending LIMELM to 54. Some of the implications were treated.
NSHELLS needed to be increased to 11 to read new data in phfit.dat
  -But then new elements needed to be added to the lists in elementnames.h
    --This included add atomic weights to dense.h. Used NIST latest values as of Aug 2025

Currently the code does not find H-like or He-like files for Ga-Xe. Files are of a different stout format .nrg.dat with is not created with NistExtractor.py.

THE CODE DOES NOT RUN CORRECTLY AS IT CAN FIND ga_31.nrg.dat

Other problems currently delayed:

DEPLETION FACTORS NOT HARDCODED ANYMORE IN abund.cpp. DEPLETION FACTORS ARE GIVEN UP TO Zn AND WITH Ge AND Kr INCLUDED IN DATA/ABUNDANCES IN 'NEW'. THESE WILL NEED TO BE REVISED AND MAKE SURE THEY ARE 0 IF THE ELEMENTS ARE NOT LISTED

I WILL MARK ALL DATA PROBLEMS IN THE CODE AS "DATAFIX NEEDED". This is to indicate a problem of missing data that might be needed to be fixed in the data files rather than in the source code.

PUSH LIMIT TO LIMELM=54
     - problem with phfit. solve (parser.cpp:1513)
       PHFIT NEEDS MORE PHOTOIONIZATION DATA
       	     -- temporary fixed using sterling_old phfit
	     THESE VALUES NEED TO BE CHECKED
     - atmdat_reading.cpp:1350 UTA files not including Ga-Kr. NEED TO FIX THAT. THE PROBLEM IS THAT THE CODE DOES NOT FIND UTA BADNELL FILES FOR Ga-Kr. NEEDS A FUTURE FIX